sudoku/sudokuvalidator_test.go
package sudoku

import (
	"fmt"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestValidate(t *testing.T) {
	testcases := []struct {
		board    [][]byte
		expected bool
	}{
		{
			board: [][]byte{
				{5, 3, 0, 0, 7, 0, 0, 0, 0},
				{6, 0, 0, 1, 9, 5, 0, 0, 0},
				{0, 9, 8, 0, 0, 0, 0, 6, 0},
				{8, 0, 0, 0, 6, 0, 0, 0, 3},
				{4, 0, 0, 8, 0, 3, 0, 0, 1},
				{7, 0, 0, 0, 2, 0, 0, 0, 6},
				{0, 6, 0, 0, 0, 0, 2, 8, 0},
				{0, 0, 0, 4, 1, 9, 0, 0, 5},
				{0, 0, 0, 0, 8, 0, 0, 7, 9},
			},
			expected: true,
		},
	}

	for _, tc := range testcases {
		t.Run(fmt.Sprintf(""), func(t *testing.T) {
			assert.Equal(t, tc.expected, Validate(tc.board))
		})
	}
}
